findFreePort() moved to tests, got rid of the portastic NPM package, createTunnel() uses random system port
