+++ package.json
@@ -44,7 +44,6 @@
     "node": ">=10"
   },
   "dependencies": {
-    "portastic": "^1.0.1",
     "underscore": "^1.9.1"
   },
   "devDependencies": {
@@ -68,6 +67,7 @@
     "isparta": "^4.0.0",
     "mocha": "^8.2.1",
     "phantomjs-prebuilt": "^2.1.16",
+    "portastic": "^1.0.1",
     "proxy": "^1.0.1",
     "request": "^2.83.0",
     "sinon": "^9.0.0",
